+++ src/components/index.ts
@@ -1,0 +1,2 @@
+import CommitElement from "./commitElement";
+export { CommitElement }
