+++ src/pages/main/index.tsx
@@ -1,5 +1,7 @@
-import { useEffect, useState } from "react";
+import { useState } from "react";
+import { selectBranches, selectRepository } from "./utils";
 import { ICommitData } from "./interfaces/commitData.interface";
+import { CommitElement } from "../../components";
 import Container from "@mui/material/Container";
 import FormControl from "@mui/material/FormControl";
 import FormHelperText from "@mui/material/FormHelperText";
@@ -7,67 +9,12 @@ import Grid from "@mui/material/Grid";
 import MenuItem from "@mui/material/MenuItem";
 import Select from "@mui/material/Select";
 import Typography from "@mui/material/Typography";
-import axios from "axios";
-import Card from "@mui/material/Card";
-import CardContent from "@mui/material/CardContent";
-import CardActions from "@mui/material/CardActions";
-import Button from "@mui/material/Button";
-import { selectBranches, selectRepository, formatDate } from "./utils";
-const CommitElement = ({ commit, nodeId }: ICommitData) => {
-  const {
-    author: { email, date, name },
-    url,
-    message,
-  } = commit;
-  return (
-    <Card elevation={2}>
-      <CardContent sx={{ m: 1 }}>
-        <Typography gutterBottom>
-          <strong>NodeId:</strong>
-        </Typography>
-        <Typography gutterBottom sx={{ fontSize: { xs: 7, md: 12 } }}>
-          {nodeId}
-        </Typography>
-        <Typography gutterBottom>
-          <strong>Author:</strong> {email}
-        </Typography>
-        <Typography gutterBottom>
-          <strong>Name:</strong> {name}
-        </Typography>
-        <Typography>
-          <strong>Date: </strong>
-          {formatDate(date)}
-        </Typography>
-        <Typography>
-          <strong>Message:</strong>
-        </Typography>
-        <Typography>{message}</Typography>
-      </CardContent>
-      <CardActions sx={{ mx: 2,pb:3 }}>
-        <Button target="_blank" href={url} variant="contained">
-          View Commit
-        </Button>
-      </CardActions>
-    </Card>
-  );
-};
+import Skeleton from "@mui/material/Skeleton";
+import useFetchData from "./hooks/useFetchData";
 const MainPage = () => {
-  const [repoSelected, setRepoSelected] = useState(selectRepository[0]);
-  const [branchSelected, setBranchSelected] = useState(selectBranches[0]);
-  const [commitsData, setCommitsData] = useState<Array<ICommitData>>();
-  useEffect(() => {
-    if (!repoSelected) return;
-    const fetchData = async () => {
-      try {
-        const url = `http://localhost:5000/commits?repository=${repoSelected}&branch=${branchSelected}`;
-        const { data } = await axios.get(url);
-        setCommitsData(data);
-      } catch (e) {
-        console.log(e);
-      }
-    };
-    fetchData();
-  }, [branchSelected, repoSelected]);
+  const [repoSelected, setRepoSelected] = useState<string>(selectRepository[0]);
+  const [branchSelected, setBranchSelected] = useState<string>(selectBranches[0]);
+  const { commitsData, error, loading } = useFetchData({ repoSelected, branchSelected })
   const handleSelectRepoChange = (value: string) => setRepoSelected(value);
   const handleSelectBranchChange = (value: string) => setBranchSelected(value);
   return (
@@ -128,7 +75,18 @@ const MainPage = () => {
             </FormHelperText>
           </FormControl>
         </Grid>
-        {commitsData &&
+        {!loading && error &&
+          <Grid item xs={12}>
+            <Typography>ERROR.....</Typography>
+          </Grid>
+        }
+        {loading &&
+          <Grid item xs={12}>
+            {[30, 100, 20].map((e) => <Skeleton variant="text" animation="wave" width={500} height={e}></Skeleton>
+            )}
+          </Grid>
+        }
+        {!loading && !error && commitsData &&
           commitsData.map((e: ICommitData, index) => (
             <Grid item xs={12} key={index}>
               <CommitElement commit={e.commit} nodeId={e.nodeId} />
